Migrate to AWS Go SDK v2 (#523)
* pkg aws ec2 api wrapper: update ec2 wrapper

* pkg aws ec2 api helper: update ec2 helper

* pkg aws ec2 api cleanup: update clean up logic

* pkg aws ec2 instance: update instance logic to use int32

* pkg utils: update util functions

* pkg provider ip eni: update package

* pkg provider branch: update package

* mocks: update mocks

* test: update tests

* update go.mod

+++ mocks/amazon-vcp-resource-controller-k8s/pkg/aws/ec2/mock_instance.go
@@ -62,7 +62,7 @@ func (mr *MockEC2InstanceMockRecorder) CurrentInstanceSecurityGroups() *gomock.C
 }
 
 // FreeDeviceIndex mocks base method.
-func (m *MockEC2Instance) FreeDeviceIndex(arg0 int64) {
+func (m *MockEC2Instance) FreeDeviceIndex(arg0 int32) {
 	m.ctrl.T.Helper()
 	m.ctrl.Call(m, "FreeDeviceIndex", arg0)
 }
@@ -89,10 +89,10 @@ func (mr *MockEC2InstanceMockRecorder) GetCustomNetworkingSpec() *gomock.Call {
 }
 
 // GetHighestUnusedDeviceIndex mocks base method.
-func (m *MockEC2Instance) GetHighestUnusedDeviceIndex() (int64, error) {
+func (m *MockEC2Instance) GetHighestUnusedDeviceIndex() (int32, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetHighestUnusedDeviceIndex")
-	ret0, _ := ret[0].(int64)
+	ret0, _ := ret[0].(int32)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
